change ip attr to float, all pitcher test pass, add views

+++ db/migrate/20180427205647_create_pitchers.rb
@@ -4,7 +4,7 @@ class CreatePitchers < ActiveRecord::Migration[5.2]
       t.string :name
       t.string :arm
       t.integer :er
-      t.integer :ip
+      t.float :ip
       t.integer :walks
       t.integer :hits
       t.integer :so
